+++ src/mods/scala/domains/terrestrial/Output.scala
@@ -2,7 +2,7 @@ package domains.terrestrial
 
 import draco._
 import io.circe.Json
-import org.apache.pekko.actor.typed.{ActorRef, Behavior, Signal, TypedActorContext}
+import org.apache.pekko.actor.typed.{Behavior, Signal, TypedActorContext}
 import org.apache.pekko.actor.typed.scaladsl.Behaviors
 
 trait Output extends DracoType
@@ -11,9 +11,9 @@ object Output extends App with DracoType {
   override lazy val typeDefinition: TypeDefinition = TypeLoader.loadType(TypeName ("Output", _namePackage = Seq ("domains", "terrestrial")))
   lazy val dracoType: Type[Output] = Type[Output] (typeDefinition)
 
-  def actorType(_consumer: => ActorRef[draco.format.json.JSON]): ActorType = new Actor[domains.world.World] {
+  def actorType(_consumer: => draco.format.json.JSON => Unit): ActorType = new Actor[domains.world.World] {
     override lazy val typeDefinition: TypeDefinition = Output.typeDefinition
-    lazy val consumer: ActorRef[draco.format.json.JSON] = _consumer
+    lazy val consumer: draco.format.json.JSON => Unit = _consumer
 
     override def receive(ctx: TypedActorContext[domains.world.World], msg: domains.world.World): Behavior[domains.world.World] = {
       msg match {
@@ -28,7 +28,7 @@ object Output extends App with DracoType {
             override lazy val typeDefinition: TypeDefinition = LocationReport.typeDefinition
             override val json: Json = payload
           }
-          consumer ! report
+          consumer(report)
         case _ => // not a Location; nothing to encode
       }
       Behaviors.same[domains.world.World]