+++ src/mods/scala/domains/world/Provider.scala
@@ -1,7 +1,7 @@
 package domains.world
 
 import draco._
-import org.apache.pekko.actor.typed.{ActorRef, Behavior, Signal, TypedActorContext}
+import org.apache.pekko.actor.typed.{Behavior, Signal, TypedActorContext}
 import org.apache.pekko.actor.typed.scaladsl.Behaviors
 
 trait Provider extends Actor[World]
@@ -10,12 +10,12 @@ object Provider extends App with DracoType {
   override lazy val typeDefinition: TypeDefinition = TypeLoader.loadType(TypeName ("Provider", _namePackage = Seq ("domains", "world")))
   lazy val dracoType: Type[Provider] = Type[Provider] (typeDefinition)
 
-  def actorType(_target: => ActorRef[World]): ActorType = new Actor[World] {
+  def actorType(_target: => World => Unit): ActorType = new Actor[World] {
     override lazy val typeDefinition: TypeDefinition = Provider.typeDefinition
-    lazy val target: ActorRef[World] = _target
+    lazy val target: World => Unit = _target
 
     override def receive(ctx: TypedActorContext[World], msg: World): Behavior[World] = {
-      target ! msg
+      target(msg)
       Behaviors.same[World]
     }